videoanaltyics-main/backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
redis==5.0.1
aioredis==2.0.1
aiohttp==3.9.1
python-dotenv==1.0.0
kafka-python==2.0.2
pillow==10.1.0
opencv-python==4.8.1.78
ultralytics==8.0.228
torch==2.1.1
torchvision==0.16.1
numpy==1.26.2
requests==2.31.0
pytest==7.4.3
pytest-asyncio==0.21.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
webrtc==0.1.0
av==10.1.0
